pyproject.toml
[tool.poetry]
name = "rediscli"
version = "0.1.0"
description = "A package using redis client"
authors = ["Tobias Newmiwaka <newmi1988@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
redis = "^3.5.3"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
